+++ packages/editor/src/app/components/overlay/modal.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { AnimatePresence, motion } from 'framer-motion';
 import * as css from './_overlay.scss';
-import { ModalProps } from '../../pages/workspace/components/overlay/overlay.types';
+import { ModalProps } from './overlay.types';
 import { Backdrop } from './backdrop';
 import { Error } from '..';
 import { menu } from '../../services';
